+++ Dependencies/CompatibilityLayers/Stress_Level_Zero_Il2Cpp/Properties/AssemblyInfo.cs
@@ -1,0 +1,13 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle(MelonLoader.BuildInfo.Description)]
+[assembly: AssemblyDescription(MelonLoader.BuildInfo.Description)]
+[assembly: AssemblyCompany(MelonLoader.BuildInfo.Company)]
+[assembly: AssemblyProduct(MelonLoader.BuildInfo.Name)]
+[assembly: AssemblyCopyright("Created by " + MelonLoader.BuildInfo.Author)]
+[assembly: AssemblyTrademark(MelonLoader.BuildInfo.Company)]
+[assembly: Guid("FEAA0159-5871-4419-9827-3CF5CAD69A53")]
+[assembly: AssemblyVersion(MelonLoader.BuildInfo.Version)]
+[assembly: AssemblyFileVersion(MelonLoader.BuildInfo.Version)]
+[assembly: MelonLoader.PatchShield]