+++ Dependencies/SupportModules/Il2Cpp/Main.cs
@@ -6,8 +6,6 @@ using UnhollowerBaseLib;
 using UnhollowerBaseLib.Runtime;
 using UnhollowerRuntimeLib;
 using UnityEngine;
-using UnityEngine.Events;
-using UnityEngine.SceneManagement;
 
 namespace MelonLoader.Support
 {
@@ -40,25 +38,7 @@ namespace MelonLoader.Support
             if (MelonLaunchOptions.Console.CleanUnityLogs)
                 ConsoleCleaner();
 
-            try
-            {
-                SceneManager.sceneLoaded = (
-                    (ReferenceEquals(SceneManager.sceneLoaded, null))
-                    ? new Action<Scene, LoadSceneMode>(OnSceneLoad)
-                    : Il2CppSystem.Delegate.Combine(SceneManager.sceneLoaded, (UnityAction<Scene, LoadSceneMode>)new Action<Scene, LoadSceneMode>(OnSceneLoad)).Cast<UnityAction<Scene, LoadSceneMode>>()
-                    );
-            }
-            catch (Exception ex) { MelonLogger.Error($"SceneManager.sceneLoaded override failed: {ex}"); }
-
-            try
-            {
-                SceneManager.sceneUnloaded = (
-                    (ReferenceEquals(SceneManager.sceneUnloaded, null))
-                    ? new Action<Scene>(OnSceneUnload)
-                    : Il2CppSystem.Delegate.Combine(SceneManager.sceneUnloaded, (UnityAction<Scene>)new Action<Scene>(OnSceneUnload)).Cast<UnityAction<Scene>>()
-                    );
-            }
-            catch (Exception ex) { MelonLogger.Error($"SceneManager.sceneUnloaded override failed: {ex}"); }
+            SceneHandler.Init();
 
             MonoEnumeratorWrapper.Register();
 
@@ -72,9 +52,6 @@ namespace MelonLoader.Support
 
             return new SupportModule_To();
         }
-
-        private static void OnSceneLoad(Scene scene, LoadSceneMode mode) { if (scene == null) return; if (MelonUtils.IsBONEWORKS) BONEWORKS_SceneHandler.OnSceneLoad(scene.buildIndex, scene.name); else Interface.OnSceneWasLoaded(scene.buildIndex, scene.name); }
-        private static void OnSceneUnload(Scene scene) { if (scene == null) return; Interface.OnSceneWasUnloaded(scene.buildIndex, scene.name); }
 
         private static Assembly Il2Cppmscorlib = null;
         private static Type streamType = null;