+++ MelonLoader/SupportModule/ISupportModule_From.cs
@@ -12,7 +12,6 @@
         void OnGUI();
         void Quit();
         void DefiniteQuit();
-        void BONEWORKS_OnLoadingScreen();
         void SetUnhollowerSupportInterface(UnhollowerSupport.Interface unhollower);
     }
 }